code/middle/search_2d_matrix/main.go
package search_2d_matrix

func searchMatrix(matrix [][]int, target int) bool {
	for i := 0; i < len(matrix); i++ {
		if target >= matrix[i][0] && target <= matrix[i][len(matrix[i])-1] {
			return binarySearch(matrix[i], target)
		}
	}
	return false
}

func binarySearch(nums []int, target int) bool {
	left, right := 0, len(nums)-1
	for left <= right {
		mid := (left + right) / 2
		if target == nums[mid] {
			return true
		} else if target < nums[mid] {
			right = mid - 1
		} else if target > nums[mid] {
			left = mid + 1
		}
	}
	return false
}
